refactor: reorganize TUI code into separate view modules
Split the 644-line app.py into three logical modules:
- app.py (279 lines): Core application, event loop, and drawing
- messages.py (270 lines): MessageView class for message management
- logs.py (147 lines): LogView class for log/session management

Also removed unused files:
- simple_app.py: Unused prototype
- ansi_app.py: Unused fallback implementation
- panes.py: Empty placeholder

This improves code organization without over-engineering. Each module
has a clear responsibility and the code is easier to maintain.

Also renamed "sessions" to "logs" throughout to match tigs terminology.
