--- python/src/tui/panes.py
@@ -1,5 +1,0 @@
-"""Pane classes for TUI - currently not used as we draw directly on stdscr."""
-
-# This file is kept for potential future use if we need to switch back to
-# a subwindow-based approach. The current implementation draws directly on
-# the main screen for better compatibility and visibility.
